+++ relay-general/src/store/normalize/span/tag_extraction.rs
@@ -2,10 +2,8 @@
 //! These are then used for metrics extraction.
 use std::collections::BTreeMap;
 
-use itertools::Itertools;
 use once_cell::sync::Lazy;
 use regex::Regex;
-use url::Url;
 
 use crate::macros::derive_fromstr_and_display;
 use crate::protocol::{Event, Span, TraceContext};
@@ -219,10 +217,9 @@ pub(crate) fn extract_tags(span: &Span, config: &Config) -> BTreeMap<SpanTagKey,
         }
 
         let domain = if span_op == "http.client" {
-            span.description
-                .value()
-                .and_then(|url| domain_from_http_url(url))
-                .map(|d| d.to_lowercase())
+            scrubbed_description
+                .and_then(|d| d.split_once(' '))
+                .map(|(_, d)| d.to_lowercase())
         } else if span_op.starts_with("db") {
             span.description
                 .value()
@@ -415,60 +412,6 @@ fn span_op_to_category(op: &str) -> Option<&str> {
         // Map everything else to unknown:
         _ => None,
     }
-}
-
-fn domain_from_http_url(url: &str) -> Option<String> {
-    match url.split_once(' ') {
-        Some((_method, url)) => {
-            let url = Url::parse(url);
-            let (domain, port) = match &url {
-                Ok(url) => (url.domain(), url.port()),
-                Err(_) => (None, None),
-            };
-            match (domain, port) {
-                (Some(domain), port) => normalize_domain(domain, port),
-                _ => None,
-            }
-        }
-        _ => None,
-    }
-}
-
-fn normalize_domain(domain: &str, port: Option<u16>) -> Option<String> {
-    if let Some(allow_listed) = normalized_domain_from_allowlist(domain, port) {
-        return Some(allow_listed);
-    }
-
-    let mut tokens = domain.rsplitn(3, '.');
-    let tld = tokens.next();
-    let domain = tokens.next();
-    let prefix = tokens.next().map(|_| "*");
-
-    let mut replaced = prefix
-        .iter()
-        .chain(domain.iter())
-        .chain(tld.iter())
-        .join(".");
-
-    if let Some(port) = port {
-        replaced = format!("{replaced}:{port}");
-    }
-
-    if replaced.is_empty() {
-        return None;
-    }
-    Some(replaced)
-}
-
-/// Allow list of domains to not get subdomains scrubbed.
-const DOMAIN_ALLOW_LIST: &[&str] = &["127.0.0.1", "localhost"];
-
-fn normalized_domain_from_allowlist(domain: &str, port: Option<u16>) -> Option<String> {
-    if let Some(domain) = DOMAIN_ALLOW_LIST.iter().find(|allowed| **allowed == domain) {
-        let with_port = port.map_or_else(|| (*domain).to_owned(), |p| format!("{}:{}", domain, p));
-        return Some(with_port);
-    }
-    None
 }
 
 #[cfg(test)]
